fix weird new code change in Emit_LoadBankPtr, fix wrong pointer conversion in EmitCodeBlock, actually do the LITTLE_ENDIAN to NUANCE_LITTLE_ENDIAN rename everywhere, make iso9660.h work for msc (not wired to UI yet)

+++ src/SuperBlock.cpp
@@ -347,8 +347,8 @@ bool SuperBlock::EmitCodeBlock(NativeCodeCache &codeCache, const bool bContainsB
       codeCache.X86Emit_MOVIR_Ptr((uintptr_t)&(codeCache.emitVars.mpe->cc), x86Reg::x86Reg_esi);
       codeCache.X86Emit_MOVIR_Ptr((uintptr_t)&(codeCache.emitVars.mpe->tempCC), x86Reg::x86Reg_edi);
 #else
-      codeCache.X86Emit_MOVIR((int32)(uintptr_t)&(codeCache.emitVars.mpe->cc), x86Reg::x86Reg_esi);
-      codeCache.X86Emit_MOVIR((int32)(uintptr_t)&(codeCache.emitVars.mpe->tempCC), x86Reg::x86Reg_edi);
+      codeCache.X86Emit_MOVIR((uintptr_t)&(codeCache.emitVars.mpe->cc), x86Reg::x86Reg_esi);
+      codeCache.X86Emit_MOVIR((uintptr_t)&(codeCache.emitVars.mpe->tempCC), x86Reg::x86Reg_edi);
 #endif
       if(bContainsBranch)
       {
